+++ src/happy_kostadin/cli.py
@@ -2,7 +2,7 @@ import argparse
 import os
 import sys
 from pathlib import Path
-from typing import TYPE_CHECKING, Dict, Any, Union
+from typing import TYPE_CHECKING, Any, Dict, Union
 
 from tqdm import tqdm
 
@@ -90,10 +90,14 @@ def main(return_checked_files: bool = False) -> Union[list, None]:
     for file in tqdm(all_files):
         content = open(file, "rb").read()
         if b"\r\n" in content:
-            files_containing_crlf.append(file)
+            file_name = str(file).replace(str(path), "")
+            files_containing_crlf.append(file_name)
 
     if files_containing_crlf:
-        raise ValueError(f"{files_containing_crlf} contains CRLF line ending.")
+        for file in files_containing_crlf:
+            print(f"Error: {file} contains CRLF line ending.")
+        print(f"Error: {len(files_containing_crlf)} files contain CRLF line ending.")
+        sys.exit(1)
     else:
         print("+=+=+ - All files are free from CRLF - +=+=+")